badarts/apis/serializers.py
from rest_framework import serializers

from homepage.models import Artist


class ArtistSerializer(serializers.ModelSerializer):
    class Meta:
        model = Artist
        fields = ['name', 'active', ]
