badarts/requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
boto3==1.19.7
botocore==1.22.7
colorclass==2.2.2
crispy-forms-materialize==0.2
dj-database-url==0.5.0
Django==4.0
django-allauth==0.47.0
django-bootstrap5==21.2
django-countries==7.2.1
django-crispy-forms==1.13.0
django-filter==21.1
django-iban-field==0.8
django-ibanfield==0.0.2
django-image-optimizer==1.0.1
django-image-uploader-widget==0.0.3
django-localflavor==3.1
django-materialize==1.0.1a2
django-multiselectfield==0.1.12
django-storages==1.12.3
djangorestframework==3.13.1
docopt==0.6.2
gunicorn==20.1.0
jmespath==0.10.0
oauthlib==3.1.1
Pillow==8.3.2
pip-upgrader==1.4.15
psycopg2==2.9.3
psycopg2-binary==2.9.1
PyJWT==2.3.0
python-resize-image==1.1.19
python-stdnum==1.16
python3-openid==3.2.0
pytz==2021.1
requests-oauthlib==1.3.0
s3transfer==0.5.0
soupsieve==2.3.1
sqlparse==0.4.2
stripe==2.60.0
terminaltables==3.1.10
tinify==1.5.2
whitenoise==6.0.0
